create music function

+++ app/src/main/java/com/longlegsdev/rhythm/presentation/screen/common/card/FavoriteMusicCard.kt
@@ -36,7 +36,7 @@ fun FavoriteMusicCard(
     title: String,
     albumImageUrl: String,
     artist: String,
-    duration: Int,
+    duration: Long,
     onFavoriteMusicClick: (Int) -> Unit = {}
 ) {
